refactor(action): move mouse apply events into src/action/terminal/
